refactor: delete adapters.py
